+++ src/owlplanner/config/schema.py
@@ -364,7 +364,6 @@ class SolverOptions(BaseModel):
     withSSTaxability: Optional[Union[str, float]] = None
     withDecomposition: Optional[str] = None
     withSSAges: Optional[Union[str, List[str]]] = None
-    withSCLoop: Optional[bool] = None
     withDuals: Optional[bool] = None
     withdrawalOrder: Optional[str] = None