+++ tests/rate_models/test_rates_reverse_roll.py
@@ -199,7 +199,6 @@ objective = "maxSpending"
 
 [solver_options]
 withMedicare = "loop"
-withSCLoop = true
 
 [results]
 default_plots = "today"